requirements.txt
Django~=4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==0.21.0
psycopg2-binary==2.9.9
drf-yasg==1.21.4
django-filter~=23.5